[PLAT-8812] Simplify Next API test setup
